requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-cors-headers==3.6.0
django-import-export==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.8
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0
